backend/requirements.txt
alembic==1.15.1
fastapi==0.115.12
pydantic~=2.11.0
pydantic-settings==2.8.1
SQLAlchemy~=2.0.40
psycopg2-binary~=2.9.9
uvicorn[standard]~=0.32.1

python-dotenv~=1.1.0
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
email-validator~=2.1.0
python-multipart~=0.0.6
bcrypt==4.0.1